backend/servicelocation/requirements.txt
asgiref==3.8.1
backports.zoneinfo==0.2.1
Django==4.2.17
django-cors-headers==4.4.0
djangorestframework==3.15.2
gunicorn==23.0.0
packaging==24.2
pillow==10.4.0
psycopg2-binary==2.9.10
sqlparse==0.5.2
typing-extensions==4.12.2
gdal<=3.6.2
